+++ tests/integration/psyoptionsAmerican.spec.ts
@@ -32,13 +32,13 @@ describe("Psyoptions American instrument integration tests", async () => {
     maker = context.maker.publicKey;
   });
 
-  it("Create buy RFQ for 1 option", async () => {
+  it("Create buy RFQ for 1 option [CALL]", async () => {
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.maker);
-    await options.mintPsyOptions(context.maker, new anchor.BN(1));
+    await options.mintPsyOptions(context.maker, new anchor.BN(1), OptionType.CALL);
 
     const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(
       context,
-      ["quote", "asset", options.callMint],
+      ["quote", "asset", options.optionMint],
       [context.taker.publicKey, context.maker.publicKey]
     );
 
@@ -65,11 +65,60 @@ describe("Psyoptions American instrument integration tests", async () => {
     // taker should receive 1 option, maker should receive 50$ and lose 1 bitcoin as option collateral
     await response.settleEscrow(maker, [taker]);
     await tokenMeasurer.expectChange([
-      { token: options.callMint, user: taker, delta: new BN(1) },
+      { token: options.optionMint, user: taker, delta: new BN(1) },
       { token: "quote", user: taker, delta: withTokenDecimals(-50) },
       { token: "quote", user: maker, delta: withTokenDecimals(50) },
       { token: "asset", user: maker, delta: withTokenDecimals(0) },
-      { token: options.callMint, user: maker, delta: new BN(-1) },
+      { token: options.optionMint, user: maker, delta: new BN(-1) },
+    ]);
+
+    await response.unlockResponseCollateral();
+    await response.cleanUp();
+  });
+
+  it("Create buy RFQ for 1 option [PUT]", async () => {
+    const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.maker, {
+      optionType: OptionType.PUT,
+    });
+    await options.mintPsyOptions(context.maker, new anchor.BN(1), OptionType.PUT);
+
+    const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(
+      context,
+      ["quote", "asset", options.optionMint],
+      [context.taker.publicKey, context.maker.publicKey]
+    );
+
+    const rfq = await context.createEscrowRfq({
+      legs: [
+        PsyoptionsAmericanInstrumentClass.create(context, options, OptionType.PUT, {
+          amount: new BN(1),
+          side: LegSide.Long,
+        }),
+      ],
+    });
+
+    // Response with agreeing to sell 2 options for 50$ or buy 5 for 45$
+    const response = await rfq.respond({
+      bid: Quote.getStandard(toAbsolutePrice(withTokenDecimals(45)), toLegMultiplier(5)),
+      ask: Quote.getStandard(toAbsolutePrice(withTokenDecimals(50)), toLegMultiplier(2)),
+    });
+
+    // Taker confirms to buy 1 option
+    await response.confirm({
+      side: QuoteSide.Ask,
+      legMultiplierBps: toLegMultiplier(1),
+    });
+    await response.prepareEscrowSettlement(AuthoritySide.Taker);
+    await response.prepareEscrowSettlement(AuthoritySide.Maker);
+
+    // taker should receive 1 option, maker should receive 50$ and lose 1 bitcoin as option collateral
+    await response.settleEscrow(maker, [taker]);
+    await tokenMeasurer.expectChange([
+      { token: options.optionMint, user: taker, delta: new BN(1) },
+      { token: "quote", user: taker, delta: withTokenDecimals(-50) },
+      { token: "quote", user: maker, delta: withTokenDecimals(50) },
+      { token: "asset", user: maker, delta: withTokenDecimals(0) },
+      { token: options.optionMint, user: maker, delta: new BN(-1) },
     ]);
 
     await response.unlockResponseCollateral();
@@ -78,10 +127,10 @@ describe("Psyoptions American instrument integration tests", async () => {
 
   it("Create sell RFQ where taker wants 2 options", async () => {
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.taker);
-    await options.mintPsyOptions(context.taker, new anchor.BN(2));
+    await options.mintPsyOptions(context.taker, new anchor.BN(2), OptionType.CALL);
     const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(
       context,
-      ["asset", "quote", options.callMint],
+      ["asset", "quote", options.optionMint],
       [taker, maker]
     );
 
@@ -99,7 +148,10 @@ describe("Psyoptions American instrument integration tests", async () => {
     const response = await rfq.respond({
       bid: Quote.getStandard(toAbsolutePrice(withTokenDecimals(45)), toLegMultiplier(2)),
     });
-    await response.confirm({ side: QuoteSide.Bid, legMultiplierBps: toLegMultiplier(2) });
+    await response.confirm({
+      side: QuoteSide.Bid,
+      legMultiplierBps: toLegMultiplier(2),
+    });
 
     // taker confirms to sell 2 options
 
@@ -119,11 +171,11 @@ describe("Psyoptions American instrument integration tests", async () => {
     await response.settleEscrow(taker, [maker]);
 
     await tokenMeasurer.expectChange([
-      { token: options.callMint, user: taker, delta: new BN(-2) },
+      { token: options.optionMint, user: taker, delta: new BN(-2) },
       { token: "quote", user: taker, delta: withTokenDecimals(90) },
       { token: "quote", user: maker, delta: withTokenDecimals(-90) },
       { token: "asset", user: maker, delta: withTokenDecimals(0) },
-      { token: options.callMint, user: maker, delta: new BN(2) },
+      { token: options.optionMint, user: maker, delta: new BN(2) },
     ]);
 
     await response.unlockResponseCollateral();
@@ -133,7 +185,7 @@ describe("Psyoptions American instrument integration tests", async () => {
   it("Create two-way RFQ with one Psyoptions American option leg, respond but maker defaults on settleEscrowment", async () => {
     // Create a two way RFQ specifying 1 option put as a leg
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.taker);
-    await options.mintPsyOptions(context.taker, new anchor.BN(2));
+    await options.mintPsyOptions(context.taker, new anchor.BN(2), OptionType.CALL);
 
     const rfq = await context.createEscrowRfq({
       activeWindow: 2,
@@ -151,11 +203,12 @@ describe("Psyoptions American instrument integration tests", async () => {
       // response with agreeing to buy 5 options for 45$
       const response = await rfq.respond({
         bid: Quote.getStandard(toAbsolutePrice(withTokenDecimals(45)), toLegMultiplier(5)),
+        expirationTimestamp: Date.now() / 1000 + 1,
       });
 
       // taker confirms to sell 2 options
       await response.confirm({ side: QuoteSide.Bid, legMultiplierBps: toLegMultiplier(2) });
-      const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(context, [options.callMint], [taker]);
+      const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(context, [options.optionMint], [taker]);
       await response.prepareEscrowSettlement(AuthoritySide.Taker);
 
       return [response, tokenMeasurer];
@@ -164,7 +217,7 @@ describe("Psyoptions American instrument integration tests", async () => {
     await response.revertEscrowSettlementPreparation(AuthoritySide.Taker);
 
     // taker have returned his assets
-    await tokenMeasurer.expectChange([{ token: options.callMint, user: taker, delta: new BN(0) }]);
+    await tokenMeasurer.expectChange([{ token: options.optionMint, user: taker, delta: new BN(0) }]);
 
     await response.settleOnePartyDefault();
     await response.cleanUp();
@@ -174,7 +227,7 @@ describe("Psyoptions American instrument integration tests", async () => {
   it("Create two-way RFQ with one Psyoptions American option leg, respond but taker defaults on settleEscrowment", async () => {
     // create a two way RFQ specifying 1 option put as a leg
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.taker);
-    await options.mintPsyOptions(context.taker, new anchor.BN(2));
+    await options.mintPsyOptions(context.taker, new anchor.BN(2), OptionType.CALL);
     const rfq = await context.createEscrowRfq({
       activeWindow: 2,
       settlingWindow: 1,
@@ -191,10 +244,14 @@ describe("Psyoptions American instrument integration tests", async () => {
       // response with agreeing to buy 5 options for 45$
       const response = await rfq.respond({
         bid: Quote.getStandard(toAbsolutePrice(withTokenDecimals(45)), toLegMultiplier(5)),
+        expirationTimestamp: Date.now() / 1000 + 1,
       });
 
       // taker confirms to sell 2 options
-      await response.confirm({ side: QuoteSide.Bid, legMultiplierBps: toLegMultiplier(2) });
+      await response.confirm({
+        side: QuoteSide.Bid,
+        legMultiplierBps: toLegMultiplier(2),
+      });
       const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(context, ["quote"], [maker]);
       await response.prepareEscrowSettlement(AuthoritySide.Maker);
       await tokenMeasurer.expectChange([{ token: "quote", user: maker, delta: withTokenDecimals(new BN(-90)) }]);
@@ -213,9 +270,13 @@ describe("Psyoptions American instrument integration tests", async () => {
 
   it("With fractional leg multiplier, rounds option amount to a bigger amount for a maker at settleEscrowment", async () => {
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.taker);
-    await options.mintPsyOptions(context.taker, new anchor.BN(1));
+    await options.mintPsyOptions(context.taker, new anchor.BN(1), OptionType.CALL);
 
-    const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(context, ["quote", options.callMint], [taker, maker]);
+    const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(
+      context,
+      ["quote", options.optionMint],
+      [taker, maker]
+    );
 
     const rfq = await context.createEscrowRfq({
       legs: [
@@ -240,8 +301,8 @@ describe("Psyoptions American instrument integration tests", async () => {
     // maker should receive 1 option(0.4 rounded up), taker should receive 40 * 0.4 = 16$
     await response.settleEscrow(taker, [maker]);
     await tokenMeasurer.expectChange([
-      { token: options.callMint, user: taker, delta: new BN(-1) },
-      { token: options.callMint, user: maker, delta: new BN(1) },
+      { token: options.optionMint, user: taker, delta: new BN(-1) },
+      { token: options.optionMint, user: maker, delta: new BN(1) },
       { token: "quote", user: taker, delta: withTokenDecimals(16) },
       { token: "quote", user: maker, delta: withTokenDecimals(-16) },
     ]);
@@ -252,9 +313,13 @@ describe("Psyoptions American instrument integration tests", async () => {
 
   it("With fractional leg multiplier, rounds option amount to a lower amount for a taker at settleEscrowment", async () => {
     const options = await AmericanPsyoptions.initalizeNewPsyoptionsAmerican(context, context.maker);
-    await options.mintPsyOptions(context.maker, new anchor.BN(2));
+    await options.mintPsyOptions(context.maker, new anchor.BN(2), OptionType.CALL);
 
-    const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(context, ["quote", options.callMint], [taker, maker]);
+    const tokenMeasurer = await TokenChangeMeasurer.takeSnapshot(
+      context,
+      ["quote", options.optionMint],
+      [taker, maker]
+    );
 
     const rfq = await context.createEscrowRfq({
       legs: [
@@ -279,8 +344,8 @@ describe("Psyoptions American instrument integration tests", async () => {
     // taker should receive 1 option(1.4 rounded down), maker should receive 50 * 1.4 = 70$
     await response.settleEscrow(maker, [taker]);
     await tokenMeasurer.expectChange([
-      { token: options.callMint, user: taker, delta: new BN(1) },
-      { token: options.callMint, user: maker, delta: new BN(-1) },
+      { token: options.optionMint, user: taker, delta: new BN(1) },
+      { token: options.optionMint, user: maker, delta: new BN(-1) },
       { token: "quote", user: taker, delta: withTokenDecimals(-70) },
       { token: "quote", user: maker, delta: withTokenDecimals(70) },
     ]);